Accepting list of states to wait
